refactor: перенести из ui
